grpc: fork headers/trailers to be separate between client/server
