+++ grpc/src/client/load_balancing/graceful_switch.rs
@@ -253,6 +253,7 @@ mod test {
     use std::sync::mpsc;
     use std::time::Duration;
 
+    use crate::client::RequestHeaders;
     use crate::client::load_balancing::ChannelController;
     use crate::client::load_balancing::LbPolicy;
     use crate::client::load_balancing::LbState;
@@ -274,7 +275,6 @@ mod test {
     use crate::client::name_resolution::Address;
     use crate::client::name_resolution::Endpoint;
     use crate::client::name_resolution::ResolverUpdate;
-    use crate::core::RequestHeaders;
     use crate::metadata::MetadataMap;
     use crate::rt::default_runtime;
