+++ grpc/src/client/load_balancing/mod.rs
@@ -32,11 +32,11 @@ use std::sync::Arc;
 use crate::StatusCodeError;
 use crate::StatusError;
 use crate::client::ConnectivityState;
+use crate::client::RequestHeaders;
 use crate::client::load_balancing::subchannel::Subchannel;
 use crate::client::load_balancing::subchannel::SubchannelState;
 use crate::client::name_resolution::Address;
 use crate::client::name_resolution::ResolverUpdate;
-use crate::core::RequestHeaders;
 use crate::metadata::MetadataMap;
 use crate::rt::GrpcRuntime;
