+++ grpc/src/client/load_balancing/test_utils.rs
@@ -31,6 +31,7 @@ use serde::Deserialize;
 use serde::Serialize;
 use tokio::sync::Notify;
 
+use crate::client::RequestHeaders;
 use crate::client::load_balancing::ChannelController;
 use crate::client::load_balancing::DynLbConfig;
 use crate::client::load_balancing::DynLbPolicy;
@@ -46,7 +47,6 @@ use crate::client::load_balancing::WorkScheduler;
 use crate::client::load_balancing::subchannel::ForwardingSubchannel;
 use crate::client::name_resolution::Address;
 use crate::client::name_resolution::ResolverUpdate;
-use crate::core::RequestHeaders;
 
 pub(crate) fn new_request_headers() -> RequestHeaders {
     RequestHeaders::default()
