+++ grpc/src/client/subchannel.rs
@@ -44,6 +44,7 @@ use crate::client::ConnectivityState;
 use crate::client::DynInvoke;
 use crate::client::DynRecvStream;
 use crate::client::DynSendStream;
+use crate::client::RequestHeaders;
 use crate::client::channel::WorkQueueItem;
 use crate::client::channel::WorkQueueTx;
 use crate::client::load_balancing::subchannel::Subchannel;
@@ -56,7 +57,6 @@ use crate::client::transport::ProxyOptions;
 use crate::client::transport::SecurityOpts;
 use crate::client::transport::TransportOptions;
 use crate::client::transport::http_connect::HttpConnectHandshaker;
-use crate::core::RequestHeaders;
 use crate::credentials::call::CallDetails;
 use crate::credentials::call::ClientConnectionSecurityInfo as CallClientConnectionSecurityInfo;
 use crate::credentials::client::ChannelSecurityInfo;
