+++ grpc/src/core/mod.rs
@@ -25,7 +25,7 @@
 //! Core gRPC types common to clients and servers.
 //!
 //! This module provides the fundamental types used in gRPC communication, such
-//! as message traits, headers, and trailers.
+//! as message traits.
 //!
 //! Most applications should not need to use these types directly, as they are
 //! typically used by generated code.  However, they may be necessary when
@@ -35,16 +35,10 @@
 //!
 //! - **[`SendMessage`] / [`RecvMessage`]:** Traits for encoding and decoding
 //!   messages.
-//! - **[`RequestHeaders`]:** Represents gRPC headers sent to the server to
-//!   initiate a request.
-//! - **[`ResponseHeaders`] / [`Trailers`]:** Represents gRPC headers and
-//!   trailers received from the server during its response.
 
 use std::any::TypeId;
 
 use bytes::Buf;
-
-use crate::metadata::MetadataMap;
 
 /// Represents a message sent by either a client or a server.
 #[allow(unused)]
@@ -116,133 +110,5 @@ impl dyn RecvMessage + '_ {
                 None
             }
         }
-    }
-}
-
-/// Contains all information transmitted in the response headers of an RPC.
-#[derive(Debug, Clone, Default)]
-pub struct ResponseHeaders {
-    metadata: MetadataMap,
-}
-
-impl ResponseHeaders {
-    /// Returns a default ResponseHeaders instance.
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Replaces the metadata of self with `metadata`.
-    pub fn with_metadata(mut self, metadata: MetadataMap) -> Self {
-        self.metadata = metadata;
-        self
-    }
-
-    /// Returns a reference to the metadata in these headers.
-    pub fn metadata(&self) -> &MetadataMap {
-        &self.metadata
-    }
-
-    /// Returns a mutable reference to the metadata in these headers.
-    pub fn metadata_mut(&mut self) -> &mut MetadataMap {
-        &mut self.metadata
-    }
-}
-
-/// Contains all information transmitted in the request headers of an RPC.
-#[derive(Debug, Clone, Default)]
-pub struct RequestHeaders {
-    /// The full (e.g. "/Service/Method") method name specified for the call.
-    method_name: String,
-    /// The application-specified metadata for the call.
-    metadata: MetadataMap,
-}
-
-impl RequestHeaders {
-    /// Returns a default RequestHeaders instance.
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Replaces the method name of self with `method_name`.
-    pub fn with_method_name(mut self, method_name: impl Into<String>) -> Self {
-        self.method_name = method_name.into();
-        self
-    }
-
-    /// Replaces the metadata of self with `metadata`.
-    pub fn with_metadata(mut self, metadata: MetadataMap) -> Self {
-        self.metadata = metadata;
-        self
-    }
-
-    /// Returns the full (e.g. "/Service/Method") method name for these headers.
-    pub fn method_name(&self) -> &String {
-        &self.method_name
-    }
-
-    /// Returns a reference to the metadata in these headers.
-    pub fn metadata(&self) -> &MetadataMap {
-        &self.metadata
-    }
-
-    /// Returns a mutable reference to the metadata in these headers.
-    pub fn metadata_mut(&mut self) -> &mut MetadataMap {
-        &mut self.metadata
-    }
-
-    /// Returns the owned fields in the RequestHeaders.
-    // TODO: make public once fields are fixed.
-    pub(crate) fn into_parts(self) -> (String, MetadataMap) {
-        (self.method_name, self.metadata)
-    }
-}
-
-/// Contains all information transmitted in the response trailers of an RPC.
-/// gRPC does not support request trailers.
-#[derive(Debug, Clone)]
-pub struct Trailers {
-    status: crate::Result<()>,
-    metadata: MetadataMap,
-}
-
-impl Trailers {
-    /// Returns a default [`Trailers`] instance.
-    pub fn new(status: crate::Result<()>) -> Self {
-        Self {
-            status,
-            metadata: MetadataMap::default(),
-        }
-    }
-
-    /// Replaces the status of self with `status`.
-    pub fn with_status(mut self, status: crate::Result<()>) -> Self {
-        self.status = status;
-        self
-    }
-
-    /// Returns a reference to the status contained in these trailers.
-    pub fn status(&self) -> &crate::Result<()> {
-        &self.status
-    }
-
-    /// Replaces the metadata of self with `metadata`.
-    pub fn with_metadata(mut self, metadata: MetadataMap) -> Self {
-        self.metadata = metadata;
-        self
-    }
-
-    /// Returns a mutable reference to the metadata in these trailers.
-    pub fn metadata_mut(&mut self) -> &mut MetadataMap {
-        &mut self.metadata
-    }
-
-    /// Returns a reference to the metadata in these trailers.
-    pub fn metadata(&self) -> &MetadataMap {
-        &self.metadata
-    }
-
-    /// Returns the status in the [`Trailers`], consuming the entire status.
-    pub fn into_status(self) -> crate::Result<()> {
-        self.status
     }
 }
