+++ grpc/src/server/mod.rs
@@ -22,6 +22,29 @@
  *
  */
 
+//! Server-side gRPC implementation and utilities.
+//!
+//! This module provides the core types and traits for building gRPC servers.
+//! While most applications will use generated code (e.g. using
+//! [`protoc-gen-rust-grpc`](https://crates.io/protoc-gen-rust-grpc)) to
+//! interact with gRPC services, this module provides the underlying primitives.
+//!
+//! # Key Concepts
+//!
+//! - **[`Server`]:** The main entry point for server-side gRPC operations.
+//! - **[`Listener`]:** A trait for accepting incoming RPCs.
+//! - **[`Handle`]:** A trait implemented by services to handle incoming RPCs.
+//!
+//! # Additional Types
+//!
+//! - **[`Call`]:** Represents an incoming RPC accepted by a [`Listener`].
+//! - **[`SendStream`] / [`RecvStream`]:** Represent the sending and receiving
+//!   sides of a server-side RPC.
+//! - **[`RequestHeaders`]:** Represents gRPC headers sent by the client to
+//!   initiate a request.
+//! - **[`ResponseHeaders`] / [`Trailers`]:** Represent gRPC headers and
+//!   trailers sent to the client during the server's response.
+
 use std::sync::Arc;
 
 use tokio::sync::oneshot;
@@ -29,10 +52,8 @@ use tonic::async_trait;
 
 use crate::client::CallOptions;
 use crate::core::RecvMessage;
-use crate::core::RequestHeaders;
-use crate::core::ResponseHeaders;
 use crate::core::SendMessage;
-use crate::core::Trailers;
+use crate::metadata::MetadataMap;
 
 pub(crate) mod interceptor;
 
@@ -265,5 +286,141 @@ impl<T: RecvStream> DynRecvStream for T {
 impl<'a> RecvStream for Box<dyn DynRecvStream + 'a> {
     async fn next(&mut self, msg: &mut dyn RecvMessage) -> Option<Result<(), ()>> {
         (**self).dyn_next(msg).await
+    }
+}
+
+/// Contains all information transmitted in the response headers of an RPC.
+#[derive(Debug, Clone, Default)]
+pub struct ResponseHeaders {
+    metadata: MetadataMap,
+}
+
+impl ResponseHeaders {
+    /// Returns a default ResponseHeaders instance.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Replaces the metadata of self with `metadata`.
+    pub fn with_metadata(mut self, metadata: MetadataMap) -> Self {
+        self.metadata = metadata;
+        self
+    }
+
+    /// Returns a reference to the metadata in these headers.
+    pub fn metadata(&self) -> &MetadataMap {
+        &self.metadata
+    }
+
+    /// Returns a mutable reference to the metadata in these headers.
+    pub fn metadata_mut(&mut self) -> &mut MetadataMap {
+        &mut self.metadata
+    }
+
+    pub(crate) fn into_metadata(self) -> MetadataMap {
+        self.metadata
+    }
+}
+
+/// Contains all information transmitted in the request headers of an RPC.
+#[derive(Debug, Clone, Default)]
+pub struct RequestHeaders {
+    /// The full (e.g. "/Service/Method") method name specified for the call.
+    method_name: String,
+    /// The application-specified metadata for the call.
+    metadata: MetadataMap,
+}
+
+impl RequestHeaders {
+    /// Returns a default RequestHeaders instance.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Replaces the method name of self with `method_name`.
+    pub fn with_method_name(mut self, method_name: impl Into<String>) -> Self {
+        self.method_name = method_name.into();
+        self
+    }
+
+    /// Replaces the metadata of self with `metadata`.
+    pub fn with_metadata(mut self, metadata: MetadataMap) -> Self {
+        self.metadata = metadata;
+        self
+    }
+
+    /// Returns the full (e.g. "/Service/Method") method name for these headers.
+    pub fn method_name(&self) -> &String {
+        &self.method_name
+    }
+
+    /// Returns a reference to the metadata in these headers.
+    pub fn metadata(&self) -> &MetadataMap {
+        &self.metadata
+    }
+
+    /// Returns a mutable reference to the metadata in these headers.
+    pub fn metadata_mut(&mut self) -> &mut MetadataMap {
+        &mut self.metadata
+    }
+
+    /// Returns the owned fields in the RequestHeaders.
+    // TODO: make public once fields are fixed.
+    pub(crate) fn into_parts(self) -> (String, MetadataMap) {
+        (self.method_name, self.metadata)
+    }
+}
+
+/// Contains all information transmitted in the response trailers of an RPC.
+/// gRPC does not support request trailers.
+#[derive(Debug, Clone)]
+pub struct Trailers {
+    status: crate::Result<()>,
+    metadata: MetadataMap,
+}
+
+impl Trailers {
+    /// Returns a default [`Trailers`] instance.
+    pub fn new(status: crate::Result<()>) -> Self {
+        Self {
+            status,
+            metadata: MetadataMap::default(),
+        }
+    }
+
+    /// Replaces the status of self with `status`.
+    pub fn with_status(mut self, status: crate::Result<()>) -> Self {
+        self.status = status;
+        self
+    }
+
+    /// Returns a reference to the status contained in these trailers.
+    pub fn status(&self) -> &crate::Result<()> {
+        &self.status
+    }
+
+    /// Replaces the metadata of self with `metadata`.
+    pub fn with_metadata(mut self, metadata: MetadataMap) -> Self {
+        self.metadata = metadata;
+        self
+    }
+
+    /// Returns a mutable reference to the metadata in these trailers.
+    pub fn metadata_mut(&mut self) -> &mut MetadataMap {
+        &mut self.metadata
+    }
+
+    /// Returns a reference to the metadata in these trailers.
+    pub fn metadata(&self) -> &MetadataMap {
+        &self.metadata
+    }
+
+    /// Returns the status in the [`Trailers`], consuming the entire status.
+    pub fn into_status(self) -> crate::Result<()> {
+        self.status
+    }
+
+    pub(crate) fn into_parts(self) -> (crate::Result<()>, MetadataMap) {
+        (self.status, self.metadata)
     }
 }
